% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicc.gllvm.R
\name{AICc}
\alias{AICc}
\alias{AICc.gllvm}
\alias{nobs}
\alias{nobs.gllvm}
\title{Corrected Akaike information criterion and number of observations}
\usage{
\method{AICc}{gllvm}(object, ...)

\method{nobs}{gllvm}(object, ...)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{...}{Not used.}
}
\description{
Calculates corrected Akaike information criterion for small sample sizes, and extracts number of observations.
}
\author{
Jenni Niku, Bert van der Veen
}
