% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFit.gllvm.R
\name{allFit.gllvm}
\alias{allFit.gllvm}
\title{Re-fit a GLLVM with different starting values or optimizers}
\usage{
\method{allFit}{gllvm}(
  object,
  seed = NULL,
  starting.vals = TRUE,
  optimizers = TRUE,
  return.best = TRUE,
  return.table = FALSE,
  sd.errors = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{seed}{seed to use for model fitting if n.init = 1.}

\item{starting.vals}{should the model be re-fitted with different starting values? Defaults to \code{TRUE}.}

\item{optimizers}{should the model be re-fitted with different optimizers? Defaults to \code{TRUE}.}

\item{return.best}{return only the best fitting object (\code{TRUE})? Or all fitted objects \code{FALSE}?}

\item{return.table}{return a table with information of the fitted model objects? Defaults to \code{FALSE}.}

\item{sd.errors}{should standard errors be returned as part of the best fitting model? Defaults to \code{TRUE}.}

\item{...}{additional arguments for a \code{\link{gllvm}} function call.}
}
\description{
The algorithm used is sensitive to the initial values, so that a single model fit might not always provide the best parameter estimates. Frequent re-fitting might be required to ensure the best solution has been found. This function refits GLLVM objects with different starting values and optimizers, for convenience, to ensure an optimal solution. 

Specifically, it implements some of the recommendations from Niku et al. 2019 with three different optimizers: 1) "zero", 2) "res" (without jitter), 3) "res3" (with jitter).
}
\examples{
\dontrun{
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y = y, family = poisson())
bestFit <- allFit(fit)
}
}
\references{
Niku, J., Brooks, W., Herliansyah, R., Hui, F. K. C., Taskinen, S., and Warton,  D. I. (2018). Efficient estimation of generalized linear latent variable models. PLoS One, 14(5):1-20.
}
\seealso{
\code{\link{gllvm}}.
}
\author{
Bert van der Veen
}
