% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.gllvm.R
\name{coef.gllvm}
\alias{coef.gllvm}
\title{Extract coefficients from gllvm}
\usage{
\method{coef}{gllvm}(object, ...)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{...}{not used.}
}
\description{
Extracts model coefficients from 'gllvm' objects.
}
\examples{
\dontrun{
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y, family = "negative.binomial")
# Coefficients
coef(fit)
}
}
\author{
David I. Warton, Jenni Niku
}
