% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.gllvm.quadratic.R
\name{confint.gllvm.quadratic}
\alias{confint.gllvm.quadratic}
\title{Confidence intervals for model parameters}
\usage{
\method{confint}{gllvm.quadratic}(object, parm = NULL, level = 0.95,
  ...)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{parm}{a specification of which parameters are to be given confidence intervals, a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level. Scalar between 0 and 1.}

\item{...\tnot}{used.}
}
\description{
Computes confidence intervals for  parameters in a fitted gllvm model.
}
\examples{
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
X <- as.matrix(antTraits$env[,1:2])
# Fit gllvm model
fit <- gllvm(y = y, X = X, family = poisson())
# 95 \% confidence intervals for coefficients of X variables
confint(fit, level = 0.95, parm = 'Xcoef')

}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>
}
