% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoCoefs.gllvm.R
\name{ecoCoefs}
\alias{ecoCoefs}
\alias{optima.gllvm}
\alias{optima}
\alias{tolerances}
\alias{tolerances.gllvm}
\title{Functions to extract ecological quantities of the latent variables from a GLLVM, if species are a quadratic function of the latent variables.}
\usage{
\method{optima}{gllvm}(object, sd.errors = TRUE, ...)

\method{tolerances}{gllvm}(object, sd.errors = TRUE, ...)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{sd.errors}{logical. If \code{TRUE}, also returns standard errors.}

\item{...}{Not used.}
}
\description{
Extracts species optima and tolerances, potentially with standard errors (derived with the Delta method).
}
\details{
Currently no separate method for calculating species maxima or gradient length are implemented.
Gradient length can be inferred from the standard deviation of the latent variables, which is reported by \code{\link{summary.gllvm}}.
}
\author{
Bert van der Veen
}
