% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.residual.cor.gllvm.R
\name{get.residual.cor.gllvm}
\alias{get.residual.cor.gllvm}
\title{Extract residual correlations from gllvm object}
\usage{
get.residual.cor.gllvm(object)
}
\arguments{
\item{object}{An object of class 'gllvm'}
}
\description{
Calculates the residual correlation from 'gllvm' models that include latent variables.
}
\examples{
\dontrun{
library(mvabund) ## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit GLLVM model
fit <- gllvm(y = y, family = "negative.binomial")
# residual correlations:
get.residual.cor.gllvm(fit)
}
}
\author{
David Warton
}
