% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnvironCov.R
\name{getEnvironCov.gllvm}
\alias{getEnvironCov.gllvm}
\alias{getEnvironCov}
\alias{getEnvironCor}
\alias{gllvm}
\title{Extract species covariances due to environmental random effects from gllvm object}
\usage{
\method{getEnvironCov}{gllvm}(object)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{...}{not used}
}
\description{
Calculates the species environment covariance matrix for a gllvm object.
}
\details{
Species covariance matrix due to the environment is calculated.

Covariances due to the covariates can only be calculated when random effects are included in the model, and are thus limited to reduced rank models (those including constrained and concurrent ordinations) fitted with random slopes, models fitted with random effects via the formula interface, or the fourth corner model fitted with random slopes.
For full rank models with random slopes, i.e., with the formula interface or the fourth corner model, the covariances of species are formulated as:

 \deqn{\Sigma_e = C*kronecker(P\rho + (1-\rho)I_p, R)*C',}
 
where P is a correlation matrix for the columns in the response (e.g., a Phylogenetic matrix), \rho the signal parameter, and R the covariance matrix for the random effects. Here, \deqn{C = kronecker(I_p, 1_r)}, with 1_r a vector of 1s equal to the number of random effects.

For reduced rank models, the covariance is separately defined for the different variance structures of the canonical coefficients in the package. With LV-specific variances, we have:

 \deqn{\Sigma_e = \Theta*S*\Theta',}

where \Theta is the matrix of loadings, and S the (diagonal) covariance matrix for the canonical coefficients. With predictor-specific variances, we instead have:

\deqn{\Sigma_e = \Sum^K_{k=1} \Theta(I_d*\sigma_k^2)\Theta',}

with I_d an identity matrix for the number of constrained and informed latent variables, and \sigma_k^2 the variance per predictor for the canonical coefficients.
}
\examples{
# Extract subset of the microbial data to be used as an example
data(microbialdata)
y <- microbialdata$Y[, order(colMeans(microbialdata$Y > 0), 
                     decreasing = TRUE)[21:40]]
fit <- gllvm(y, family = poisson())
fit$logL
cr <- getResidualCor(fit)

}
\seealso{
\code{\link{getEnvironCor}} ,\code{\link{getResidualCov.gllvm}}, \code{\link{getResidualCor.gllvm}},.
}
\author{
Bert van der Veen
}
