% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoadings.gllvm.R
\name{getLoadings.gllvm}
\alias{getLoadings.gllvm}
\alias{getLoadings}
\title{Extract loadings}
\usage{
\method{getLoadings}{gllvm}(object, ...)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{...}{not used}
}
\description{
Extract loadings (species scores) from a gllvm object.
}
\details{
Function retrieves the loadings a.k.a. species scores for a GLLVM. For the optima of a quadratic response model, see \code{\link{optima.gllvm}}
}
