% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLV.gllvm.quadratic.R
\name{getLV.gllvm.quadratic}
\alias{getLV.gllvm.quadratic}
\alias{getLV}
\title{Extract latent variables}
\usage{
\method{getLV}{gllvm.quadratic}(object)
}
\arguments{
\item{object}{an object of class 'gllvm'.}
}
\description{
Extract latent variables from gllvm object.
}
