% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResidualCor.gllvm.R
\name{getResidualCor.gllvm}
\alias{getResidualCor.gllvm}
\alias{getResidualCor}
\title{Extract residual correlations from gllvm object}
\usage{
\method{getResidualCor}{gllvm}(object, adjust = 1)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{adjust}{defaults to 1, when residual covariance is adjusted in the case of negative binomial and binomial distribution. Alternatives are 0, when adjustment is not used, and 2 for negative binomial distribution. See function \code{\link{getResidualCov.gllvm}}.}
}
\description{
Calculates the residual correlation matrix for gllvm model.
}
\examples{
# Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y = y, family = poisson())
# residual correlations:
cr <- getResidualCor(fit)
\dontrun{
# Plot residual correlations:
install.packages("corrplot", "gclus")
library(corrplot)
library(gclus)
corrplot(cr[order.single(cr), order.single(cr)], diag = F,
  type = "lower", method = "square", tl.cex = 0.8, tl.srt = 45, tl.col = "red")
  }

}
\author{
Francis K.C. Hui, Jenni Niku, David I. Warton
}
