% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gllvm.quadratic.R
\name{gllvm.quadratic}
\alias{gllvm.quadratic}
\title{Generalized Linear Latent Variable Models}
\usage{
gllvm.quadratic(
  y = NULL,
  X = NULL,
  TR = NULL,
  data = NULL,
  formula = NULL,
  num.lv = 2,
  family,
  row.eff = FALSE,
  offset = NULL,
  sd.errors = TRUE,
  Lambda.struc = "unstructured",
  diag.iter = 0,
  trace = FALSE,
  trace2 = FALSE,
  n.init = 1,
  reltol = 1e-08,
  seed = NULL,
  maxit = 10000,
  start.fit = NULL,
  starting.val = "res",
  optimizer = "optim",
  Lambda.start = c(0.1, 0.5),
  jitter.var = 0,
  ridge = FALSE,
  ridge.quadratic = FALSE,
  par.scale = 1,
  fn.scale = 1,
  grad.check = FALSE,
  zeta.struc = "species",
  maxit.lingllvm = NULL,
  starting.val.lingllvm = "res",
  common.tolerances = FALSE,
  parallel = FALSE,
  start.struc = "common",
  gamma1 = 0,
  gamma2 = 0,
  theta4 = NULL,
  Lambda2.start = 0.01
)
}
\arguments{
\item{y}{(n x m) matrix of responses.}

\item{X}{matrix or data.frame of environmental covariates.}

\item{TR}{matrix or data.frame of trait covariates.}

\item{data}{data in long format, that is, matrix of responses, environmental and trait covariates and row index named as ?\200\231id?\200\231. When used, model needs to be defined using formula. This is alternative data input for y, X and TR.}

\item{formula}{an object of class 'formula' (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{num.lv}{number of latent variables, d, in gllvm model. Non-negative integer, less than number of response variables (m). Defaults to 2.}

\item{family}{distribution function for responses. Options are \code{poisson(link = 'log')}, \code{'negative.binomial'} (with log link), \code{binomial(link = 'probit')} and \code{'ordinal'}.}

\item{row.eff}{\code{FALSE}, \code{TRUE} or \code{'random'}, Indicating whether row effects are included in the model as a fixed or as a random effects. Defaults to \code{FALSE} when row effects are not included.}

\item{offset}{vector or matrix of offset terms.}

\item{sd.errors}{logical. If \code{TRUE} (default) standard errors for parameter estimates are calculated.}

\item{Lambda.struc}{covariance structure of VA distributions for latent variables,'unstructured' or 'diagonal'.}

\item{diag.iter}{if larger than 1, the updating of variational (covariance) parameters is sped-up by first fitting a model with a diagonal structure. If two values are supplied, uses the first for the linear GLLVM and second for the quadratic GLLVM, if \code{starting.val="lingllvm"}}

\item{trace}{logical, if \code{TRUE} in each iteration step of n.init information on current step will be printed. Defaults to FALSE.}

\item{trace2}{local, if \code{TRUE} prints the optimizer trace.}

\item{n.init}{number of initial runs. Uses multiple runs and picks up the one giving highest log-likelihood value. Defaults to 10. If \code{starting.val="lingllvm"} and \code{starting.val.lingllvm=="res"}, it is used for the linear GLLVM instead.}

\item{reltol}{convergence criteria for log-likelihood, defaults to 1e-6.}

\item{seed}{a single seed value, defaults to \code{NULL}.}

\item{maxit}{maximum number of iterations within \code{optim} function, defaults to 1000.}

\item{start.fit}{object of class 'gllvm' or 'gllvm.quadratic' which can be given as starting parameters. When an object of class 'gllvm' is provided, starting values for quadratic coefficients are generated with chosen 'starting.val'.}

\item{starting.val}{starting values can be generated by fitting a linear GLLVM, "lingllvm". Alternatively by the same procedure as for the linear GLLVM: fitting model without latent variables, and applying factorial analysis to residuals to get starting values for latent variables and their coefficients (\code{starting.val = 'res'}). Another options are to use zeros as a starting values (\code{starting.val = 'zero'}) or initialize starting values for latent variables with (n x num.lv) matrix. Defaults to \code{'gllvm'}, which is recommended.}

\item{optimizer}{The log-likelihood can be optimized using \code{'\link{optim}'} (default) or \code{'\link{nlminb}'} (only with diag.iter=0).}

\item{Lambda.start}{starting values for variances in VA distributions for latent variables in variational approximation method. Defaults to 0.1.}

\item{jitter.var}{jitter variance for starting values of latent variables. Defaults to 0, meaning no jittering.}

\item{par.scale}{(see \code{'\link{optim}'} for details) Here, either "coef" to scale all parameters to equal magnitudes, or a single value to scale all parameters with, e.g. 0.0001)}

\item{fn.scale}{(see \code{'\link{optim}'} for details)}

\item{grad.check}{defaults to TRUE. Checks if TMB provided gradient is near zero, i.e. if the model has converged.}

\item{zeta.struc}{Structure for cut-offs in the ordinal model. Either "common", for the same cut-offs for all species, or "species" for species-specific cut-offs. For the latter, classes are arbitrary per species, each category per species needs to have at least one observations. Defaults to "species".}

\item{starting.val.lingllvm}{Procedure to generate starting values for linear GLLVM if \code{starting.val="lingllvm"}. See also \link{gllvm}.}

\item{common.tolerances}{whether to fit quadratic model with equal tolerances (one quadratic coefficient per latent variable) or species-specific tolerances.}

\item{parallel}{whether to turn of support for parallel computing. Note that either \link{openmp} must be used if \code{n.init=1}. Or a parallel back-end supported by \link{foreach}.}

\item{Lambda2.start}{starting values for the quadratic coefficients (should be a scalar). Defaults to 0.01.}
}
\value{
An object of class 'gllvm' includes the following components:


 \item{call }{function call}
 \item{logL }{log likelihood}
 \item{lvs }{latent variables}
 \item{params}{list of parameters
 \itemize{
   \item{theta }{ coefficients related to latent variables}
   \item{beta0 }{ column specific intercepts}
   \item{Xcoef }{ coefficients related to environmental covariates X}
   \item{B }{ coefficients in fourth corner model}
   \item{row.params }{ row-specific intercepts}
   \item{phi }{ dispersion parameters \eqn{\phi} for negative binomial or standard deviation for gaussian family}
   \item{inv.phi }{ dispersion parameters \eqn{1/\phi} for negative binomial}
   }}
 \item{sd }{ list of standard errors of parameters}
 \item{prediction.errors }{ list of prediction covariances for latent variables and variances for random row effects when method \code{'LA'} is used}
 \item{A, Ar }{ covariance matrices for variational densities of latent variables and variances for random row effects}
}
\description{
Fits a generalized linear latent variable model for multivariate data, with quadratic latent variables. The model is fitted with variational
approximations.
}
\details{
Fits the species packing model by generalized linear latent variable models with quadratic latent variables.
Method can be used with two types of latent variable models depending on covariates. If only
site related environmental covariates are used, the expectation of response \eqn{Y_{ij}} is determined by

\deqn{g(\mu_{ij}) = \eta_{ij} = \alpha_i + \beta_{0j} + x_i'\beta_j + u_i'\theta_j + u_i'\D_{j}u_i,}

where \eqn{g(.)} is a known link function, \eqn{u_i} are \eqn{d}-variate latent variables (\eqn{d}<<\eqn{m}), \eqn{\alpha_i} is an optional row effect
at site \eqn{i}, and it can be fixed or random effect, \eqn{\beta_{0j}} is an intercept term for species \eqn{j}, \eqn{\beta_j}, \eqn{\theta_j} and \eqn{D_j}  are column
specific coefficients related to covariates and the latent variables, respectively.

An alternative model is the fourth corner model (Brown et al., 2014, Warton et al., 2015) which will be fitted if also trait covariates
are included. The expectation of response \eqn{Y_{ij}} is

\deqn{g(\mu_{ij}) = \alpha_i + \beta_{0j} + x_i'\beta_x + TR_j'\beta_t + vec(B)*kronecker(TR_j,X_i) + u_i'\theta_j + u_i'\D_{j}u_i}

where g(.), \eqn{u_i}, \eqn{\beta_{0j}} and \eqn{\theta_j} are defined as above. Vectors \eqn{\beta_x} and \eqn{\beta_t} are the main effects
or coefficients related to environmental and trait covariates, respectively, matrix \eqn{B} includes interaction terms.
The interaction/fourth corner terms are optional as well as are the main effects of trait covariates.


The method is sensitive for the choices of initial values of the latent variables. Therefore it is
recommendable to use multiple runs and pick up the one giving the highest log-likelihood value.
However, sometimes this is computationally too demanding, and default option
\code{starting.val = 'res'} is recommended. For more details on different starting value methods, see Niku et al., (2018).

Models are implemented using TMB (Kristensen et al., 2015) applied to variational approximation (Hui et al., 2017).

With ordinal family response classes must start from 0 or 1.

\subsection{Distributions}{

  Mean and variance for distributions are defined as follows.
\itemize{
  \item{For count data \code{family = poisson()}:} {Expectation \eqn{E[Y_{ij}] = \mu_{ij}}, variance \eqn{V(\mu_{ij}) = \mu_{ij}}, or}
  \item{ \code{family = 'negative.binomial'}:}{ Expectation \eqn{E[Y_{ij}] = \mu_{ij}}, variance \eqn{V(\mu_{ij}) = \mu_{ij}+\frac{\mu_{ij}}{\phi_j}}}

  \item{For binary data \code{family = binomial()}:}{ Expectation \eqn{E[Y_{ij}] = \mu_{ij}}, variance \eqn{V(\mu_{ij}) = \mu_{ij}(1-\mu_{ij})}.}

  \item{For ordinal data \code{family = 'ordinal'}:}{ Cumulative probit model, see Hui et.al. (2016).}
}
}
}
\note{
If function gives warning: 'In f(x, order = 0) : value out of range in 'lgamma'', optimizer have visited an area where gradients become too big. It is automatically fixed by trying another step in the optimization process, and can be ignored if errors do not occur.
}
\examples{
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
X <- as.matrix(antTraits$env)
TR <- antTraits$traits
# Fit model with environmental covariates Bare.ground and Shrub.cover
fit <- gllvm(y, X,
  formula = ~ Bare.ground + Shrub.cover,
  family = poisson()
)
ordiplot(fit)
coefplot(fit)
\donttest{
## Example 1: Fit model with two latent variables
# Using variational approximation:
fitv0 <- gllvm(y, family = "negative.binomial", method = "VA")
ordiplot(fitv0)
plot(fitv0, mfrow = c(2, 2))
summary(fitv0)
confint(fitv0)

## Example 2: gllvm with environmental variables
# Fit model with two latent variables and all environmental covariates,
fitvX <- gllvm(formula = y ~ X, family = "negative.binomial")
ordiplot(fitvX, biplot = TRUE)
coefplot(fitvX)
# Fit model with environmental covariates Bare.ground and Shrub.cover
fitvX2 <- gllvm(y, X,
  formula = ~ Bare.ground + Shrub.cover,
  family = "negative.binomial"
)
ordiplot(fitvX2)
coefplot(fitvX2)
# Use 5 initial runs and pick the best one
fitvX_5 <- gllvm(y, X,
  formula = ~ Bare.ground + Shrub.cover,
  family = "negative.binomial", n.init = 5, jitter.var = 0.1
)
ordiplot(fitvX_5)
coefplot(fitvX_5)

## Example 3: Data in long format
# Reshape data to long format:
datalong <- reshape(data.frame(cbind(y, X)),
  direction = "long",
  varying = colnames(y), v.names = "y"
)
head(datalong)
fitvLong <- gllvm(
  data = datalong, formula = y ~ Bare.ground + Shrub.cover,
  family = "negative.binomial"
)

## Example 4: Fourth corner model
# Fit fourth corner model with two latent variables
fitF1 <- gllvm(y = y, X = X, TR = TR, family = "negative.binomial")
coefplot(fitF1)
# Fourth corner can be plotted also with next lines
# fourth = fitF1$fourth.corner
# library(lattice)
# a = max( abs(fourth) )
# colort = colorRampPalette(c('blue','white','red'))
# plot.4th = levelplot(t(as.matrix(fourth)), xlab = 'Environmental Variables',
#              ylab = 'Species traits', col.regions = colort(100),
#              at = seq( -a, a, length = 100), scales = list( x = list(rot = 45)))
# print(plot.4th)

# Specify model using formula
fitF2 <- gllvm(
  y = y, X = X, TR = TR,
  formula = ~ Bare.ground + Canopy.cover * (Pilosity + Webers.length),
  family = "negative.binomial"
)
ordiplot(fitF2)
coefplot(fitF2)

## Example 5: Random row effects
fitRand <- gllvm(y, family = "negative.binomial", row.eff = "random")
ordiplot(fitRand, biplot = TRUE)
}
}
\references{
Brown, A. M., Warton, D. I., Andrew, N. R., Binns, M., Cassis, G., and Gibb, H. (2014). The fourth-corner solution - using predictive models to understand how species traits interact with the environment. Methods in Ecology and Evolution, 5:344-352.

Dunn, P. K. and Smyth, G. K. (2005).  Series evaluation of tweedie exponential dispersion model densities. Statistics and Computing, 15:267-280.

Hui, F. K. C., Taskinen, S., Pledger, S., Foster, S. D., and Warton, D. I. (2015).  Model-based approaches to unconstrained ordination. Methods in Ecology and Evolution, 6:399-411.

Hui, F. K. C., Warton, D., Ormerod, J., Haapaniemi, V., and Taskinen, S. (2017).  Variational approximations for generalized linear latent variable models. Journal of Computational and Graphical Statistics. Journal of Computational and Graphical Statistics, 26:35-43.

Kasper Kristensen, Anders Nielsen, Casper W. Berg, Hans Skaug, Bradley M. Bell (2016). TMB: Automatic Differentiation and Laplace Approximation. Journal of Statistical Software, 70(5), 1-21.

Niku, J., Warton,  D. I., Hui, F. K. C., and Taskinen, S. (2017). Generalized linear latent variable models for multivariate count and biomass data in ecology. Journal of Agricultural, Biological, and Environmental Statistics, 22:498-522

Niku, J., Brooks, W., Herliansyah, R., Hui, F. K. C., Taskinen, S., and Warton,  D. I. (2018). Efficient estimation of generalized linear latent variable models. Submitted.

Warton, D. I., Guillaume Blanchet, F., O'Hara, R. B., Ovaskainen, O., Taskinen, S., Walker, S. C. and Hui, F. K. C. (2015). So many variables: Joint modeling in community ecology. Trends in Ecology & Evolution, 30:766-779.
}
\seealso{
\code{\link{coefplot.gllvm.quadratic}}, \code{\link{confint.gllvm.quadratic}}, \code{\link{ordiplot.gllvm.quadratic}}, \code{\link{plot.gllvm.quadratic}}, \code{\link{residuals.gllvm}}, \code{\link{summary.gllvm.quadratic}}.
}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>, Wesley Brooks, Riki Herliansyah, Francis K.C. Hui, Sara Taskinen, David I. Warton
}
