% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.gllvm.quadratic.R
\name{logLik.gllvm.quadratic}
\alias{logLik.gllvm.quadratic}
\title{Log-likelihood of gllvm}
\usage{
\method{logLik}{gllvm.quadratic}(object, ...)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{...}{not used.}
}
\description{
Extracts Log-likelihood from 'gllvm' objects.
}
\examples{
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y = y, family = poisson())
# log-Likelihood:
logLik(fit)
}
\author{
David I. Warton, Jenni Niku
}
