% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvplot.gllvm.quadratic.R
\name{lvplot.gllvm.quadratic}
\alias{lvplot.gllvm.quadratic}
\alias{lvplot}
\title{Plot covariate coefficients and confidence intervals}
\usage{
\method{lvplot}{gllvm.quadratic}(
  object,
  plot.optima = TRUE,
  y.label = TRUE,
  which.lvs = NULL,
  cex.ylab = 0.5,
  mfrow = NULL,
  mar = c(4, 6, 2, 1),
  xlim.list = rep(list(c(-5, 5)), length(which.lvs)),
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{plot.optima}{logical, defaults to \code{TRUE}. If \code{FALSE}, only plots species tolerances.}

\item{y.label}{logical, if \code{TRUE} (default) colnames of y with respect to coefficients are added to plot.}

\item{which.lvs}{vector indicating which LVs will be plotted. Default is \code{NULL} when optima and tolerances of all LVs are plotted.}

\item{cex.ylab}{the magnification to be used for axis annotation relative to the current setting of cex.}

\item{mfrow}{same as \code{mfrow} in \code{par}. If \code{NULL} (default) it is determined automatically.}

\item{mar}{vector of length 4, which defines the margin sizes: \code{c(bottom, left, top, right)}. Defaults to \code{c(4,5,2,1)}.}

\item{xlim.list}{list of vectors with length of two to define the intervals for an x axis in each covariate plot. Defaults to NULL when the interval is defined by the range of point estimates and confidence intervals}

\item{level}{the confidence level. Scalar between 0 and 1.}

\item{...}{additional graphical arguments.}
}
\description{
Plots covariate coefficients and their confidence intervals.
}
\details{
Species optima and tolerances per latent variable are plotted with confidence intervals. Ablines at -2,2 and -5,5 are added for assistance, but do not have any meaning. Tolerances of which the CI of the quadratic coefficient crosses zero are greyed out. Tolerances are shown opposite of optima for readability, though are always positive. Optima of which the SE is larger than the optima are greyed out. Statistical uncertainties of optima that are larger than  15 or smaller than -15 are not plotted by default.
}
\author{
Bert van der Veen
}
