% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche.overlap.R
\name{niche.overlap.gllvm.quadratic}
\alias{niche.overlap.gllvm.quadratic}
\alias{niche.overlap}
\title{Niche overlap}
\usage{
\method{niche.overlap}{gllvm.quadratic}(object, which.lvs = 1,
  spp = NULL, option = 1)
}
\arguments{
\item{object}{An object of class 'gllvm.quadratic'.}

\item{which.lvs}{The latent variables niche overlap should be calculated for.}

\item{spp}{Species index, if specified calculates the probability of species occuring jointly.}

\item{option}{either 1 or 2. If 1, calculates the matrix of pairwise occurence probabilities. If 2 calculates the niche overlap coefficient as specified in Mutshinda and O' Hara (2011). Essentially, the probability of joint occurrence for species i and j, relative to the probability of the niche belonging to species i.}
}
\description{
Calculates the overlap between niches of species
}
\details{
The niche overlap is calculated per latent variable. If multiple latent variables are selected, calculates the average niche overlap.
}
\examples{
# Load a dataset from the mvabund package
data(spider)
# Fit quadratic gllvm
fit <- gllvm.quadratic(y = y, family = poisson(),num.lv=2)
# calculate the niche overlap for the first latent variable
overlap <- niche.overlap(fit, which.lvs = 1)

# plot results
library(corrplot)
par(mfrow=c(1,2))
corrplot(overlap, cl.lim = c(0,1))
ordiplot(fit,which.lvs=1)

# plot residual correlations of species with >20\% niche overlap in two dimensions
corrplot(gllvm.quadratic::getResidualCor(test),p.mat=1-niche.overlap(test,which.lvs=1:2),sig.level=.8)

# calculate probability of multiple species occuring together in two dimensions
niche.overlap(test,which.lvs=1:2,spp=9:12)

# calculate probability of a species occurring in the combined latent space
niche.overlap(test,which.lvs=1:2,spp=1)
}
\references{
Inman, H.F., Bradley Jr, E.L. (1989). The overlapping coefficient as a measure of agreement between probability distributions and point estimation of the overlap of two normal densities. Communications in Statistics - Theory and methods, 18:3851-3874.
Mutshinda C.M., O'Hara, R.B. (2011) Integrating the niche and neutral perspectives on community structure and dynamics. Oecologia 166: 241-251.
}
\author{
Bert van der Veen
}
