% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoCoefs.gllvm.R
\name{optima}
\alias{optima}
\alias{optima.gllvm}
\alias{tolerances}
\alias{tolerances.gllvm}
\alias{gradient.length}
\alias{gradient.length.gllvm}
\title{Functions to extract ecological quantities of the latent variables from a GLLVM, if species are a quadratic function of the latent variables.}
\usage{
\method{optima}{gllvm}(object, sd.errors = TRUE)

\method{tolerances}{gllvm}(object, sd.errors = TRUE)

\method{gradient.length}{gllvm}(object, sd.errors = TRUE)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{sd.errors}{logical. If \code{TRUE}, also returns standard errors.}
}
\description{
Extracts species optima, tolerances or gradient lengths, potentially with standard errors (derived with the Delta method).
}
\author{
Bert van der Veen
}
