% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optiplot.gllvm.quadratic.R
\name{optiplot.gllvm.quadratic}
\alias{optiplot.gllvm.quadratic}
\alias{optiplot}
\title{Plot species optima for latent variables from a gllvm}
\usage{
\method{optiplot}{gllvm.quadratic}(
  object,
  ind.spp = NULL,
  alpha = 0.5,
  main = NULL,
  which.lvs = NULL,
  s.colors = 1,
  s.labels = "rug",
  cex.spp = 0.7,
  opt.region = ifelse(length(which.lvs) == 1, "confidence", "distribution"),
  type = "response",
  intercept = TRUE,
  legend = FALSE,
  scale = FALSE,
  site.region = FALSE,
  lty.ellips = c("solid", "dashed"),
  lwd.ellips = 1,
  col.ellips = "gray",
  alpha.col = 0.4,
  level = 0.95,
  ylim = NULL,
  xlim = NULL,
  length.out = 100,
  ...
)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{ind.spp}{the number of response variables (usually, species) to include on the biplot (sorted by varianve explained). The default is all. Can supply a vector, in which case those species will be picked from the order of variation explained on the latent variable (e.g. the species that explain fourth and sixth most variation).}

\item{alpha}{a numeric scalar between 0 and 1 that is used to control the relative scaling of the latent variables and their coefficients, when constructing a biplot.}

\item{main}{main title.}

\item{which.lvs}{indices of two latent variables to be plotted if number of the latent variables is more than 2. A vector with length of two. Defaults to \code{c(1,2)}.}

\item{s.colors}{colors for sites}

\item{s.labels}{Can be FALSE, for no labels, "row.names" or "rug".}

\item{cex.spp}{size of species labels in biplot}

\item{opt.region}{Only for 2D plots, efaults to FALSE. If "confidence", plots confidence intervals for species optima as ellipses. If "distribution" plots predicted species distributions.}

\item{type}{Can be used to predict on the response or link scale. Default is response (except for ordinal, for which the only option is "link").}

\item{intercept}{Can be used to include species-intercepts in the plot. Default is TRUE}

\item{legend}{when \code{TRUE} adds legend in the topleft corner of the plot, instead of species names in the plot}

\item{scale}{For 2D plots, either "FALSE",species" or "sites" to scale optima or site scores by the ratio variance explained. Alternatively can be "tolerances" to scale optima by tolerances and site scores by average tolerances per latent variable.}

\item{site.region}{logical, if \code{TRUE} prediction regions for the predicted latent variables are plotted, defaults to \code{FALSE}.}

\item{lty.ellips}{line type for prediction ellipses. Can be a vector of size two. See graphical parameter lty.}

\item{lwd.ellips}{line width for prediction ellipses. Can be a vector of size two. See graphical parameter lwd.}

\item{col.ellips}{colors for site prediction ellipses.}

\item{alpha.col}{used to control the transparency of confidence interval ribbons in 1D plot, and prediction regions in 2D plot.}

\item{level}{level for prediction regions. Can be a vector of size two. Defaults to 0.95.}

\item{length.out}{number of points to evaluate the function at for 1D plots. Defaults to 100.}

\item{...}{additional graphical arguments.}
}
\description{
Plots latent variables and their corresponding species optima.
}
\details{
Function constructs a plot of species optima. in 1D a dashed line indicates the optimum of species. In 2D species distributions can be added as ellipses.
}
\examples{
#' ## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
fit <- gllvm.quadratic(y, family = poisson())
# Ordination plot:
optiplot(fit)
# Optiplot with 10 species
optiplot(fit, ind.spp = 10)
}
\author{
Bert van der Veen
}
