% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiplot.gllvm.quadratic.R
\name{ordiplot.gllvm.quadratic}
\alias{ordiplot.gllvm.quadratic}
\alias{ordiplot}
\title{Plot latent variables from gllvm model}
\usage{
\method{ordiplot}{gllvm.quadratic}(object, biplot = FALSE,
  ind.spp = NULL, alpha = 0.5, main = NULL, which.lvs = NULL,
  jitter = FALSE, jitter.amount = 0.2, s.colors = 1,
  symbols = FALSE, cex.spp = 0.7, bell = TRUE, env.ranges = FALSE,
  type = "response", intercept = TRUE, legend = FALSE,
  scale = "species", predict.region = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{biplot}{\code{TRUE} if both latent variables and their coefficients are plotted, \code{FALSE} if only latent variables.}

\item{ind.spp}{the number of response variables (usually, species) to include on the biplot. The default is none, or all if \code{biplot = TRUE}.}

\item{alpha}{a numeric scalar between 0 and 1 that is used to control the relative scaling of the latent variables and their coefficients, when constructing a biplot.}

\item{main}{main title.}

\item{which.lvs}{indices of two latent variables to be plotted if number of the latent variables is more than 2. A vector with length of two. Defaults to \code{c(1,2)}.}

\item{jitter}{if \code{TRUE}, jittering is applied on points.}

\item{jitter.amount}{numeric, positive value indicating an amount of jittering for each point, defaults to 0.2 (jitter range).}

\item{s.colors}{colors for sites}

\item{symbols}{logical, if \code{TRUE} sites are plotted using symbols, if \code{FALSE} (default) site numbers are used}

\item{cex.spp}{size of species labels in biplot}

\item{bell}{logical, if TRUE plots bell-shapes (1D) or biplot with (scaled) predicted optima and 95 percent of the predicted environmental ranges (2D)}

\item{env.ranges}{logical, if TRUE plots predicted species distributions in 2D}

\item{type}{when predicting bell-shapes, can be used to predict on the response or link scale. Default is response (except for ordinal, for which the only option is "link").}

\item{intercept}{when predicting bell-shapes, can be used to include species-intercepts in the plot. Default is TRUE}

\item{legend}{when \code{TRUE} adds legend in the topleft corner of the plot, instead of species names in the plot}

\item{scale}{For 2D plots, either "species" or "sites" to scale optima or site scores by the ratio variance explained. Alternatively can be "tolerances" to scale optima by tolerances and site scores by average tolerances per latent variable.}

\item{predict.region}{logical, if \code{TRUE} prediction regions for the predicted latent variables are plotted, defaults to \code{FALSE}.}

\item{leve}{level for prediction regions.}

\item{...\tadditional}{graphical arguments.}
}
\description{
Plots latent variables and their corresponding coefficients (biplot).
}
\details{
Function constructs a scatter plot of two latent variables, i.e. an ordination plot. If only one latent
variable is in the fitted model, latent variables are plotted against their row indices \code{bell=F}.
or if \code{bell=T} against their predictions.

Coefficients related to latent variables are plotted in the same figure with the latent
variables if \code{biplot = TRUE, bell = F}. They are labeled using the column names of y. The number
of latent variable coefficients to be plotted can be controlled by ind.spp. An argument alpha
is used to control the relative scaling of the latent variables and their coefficients.
If \code{alpha = 0.5}, the latent variables and their coefficients are on the same scale.

If \code{bell=T} and two latent variables are selected (which.lvs), species optima are plotted with predicted environmental ranges.
}
\examples{
#'## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
fit <- gllvm(y, family = poisson())
# Ordination plot:
ordiplot(fit)
# Biplot with 10 species
ordiplot(fit, biplot = TRUE, ind.spp = 10)

}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>, Francis K.C. Hui
}
