% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiplot.gllvm.R
\name{ordiplot}
\alias{ordiplot}
\title{Plot latent variables from gllvm model}
\usage{
ordiplot(object, ...)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{...}{additional graphical arguments.}

\item{biplot}{\code{TRUE} if both latent variables and their coefficients are plotted, \code{FALSE} if only LVs.}

\item{ind.spp}{the number of response variables (usually, species) to include on the biplot. The default is none, or all if \code{biplot = TRUE}.}

\item{alpha}{a numeric scalar between 0 and 1 that is used to control the relative scaling of the latent variables and their coefficients, when constructing a biplot.}

\item{main}{main title.}

\item{which.lvs}{indices of two latent variables to be plotted if number of the latent variables is more than 2. A vector with length of two. Defaults to \code{c(1,2)}.}

\item{jitter}{if \code{TRUE}, jittering is applied on points.}

\item{s.colors}{colors for sites}

\item{symbols}{logical, if \code{TRUE} sites are plotted using symbols, if \code{FALSE} (default) site numbers are used}
}
\description{
Plots latent variables and their corresponding coefficients (biplot).
}
\details{
Function constructs a scatter plot of two latent variables, i.e. an ordination plot. If only one latent
variable is in the fitted model, latent variables are plotted against their corresponding row indices.
The latent variables are labeled using the row index of the response matrix y.

Coefficients related to latent variables are plotted in the same figure with the latent
variables if \code{biplot = TRUE}. They are labeled using the column names of y. The number
of latent variable coefficients to be plotted can be controlled by ind.spp. An argument alpha
is used to control the relative scaling of the latent variables and their coefficients.
If \code{alpha = 0.5}, the latent variables and their coefficients are on the same scale.
}
\examples{
#'## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
fit <- gllvm(y, family = "poisson")
# Ordination plot:
ordiplot.gllvm(fit)
# Biplot with 10 species
ordiplot.gllvm(fit, biplot = TRUE, ind.spp = 10)

}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>, Francis K.C. Hui
}
