% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordplot.gllvm.R
\name{ordplot.gllvm}
\alias{ordplot.gllvm}
\title{Plot latent variables from GLLVM model.}
\usage{
ordplot.gllvm(object, biplot = FALSE, ind.spp = NULL, alpha = 0.5,
  main = NULL, which.lvs = c(1, 2), jitter = FALSE, ...)
}
\arguments{
\item{object}{An object of class 'gllvm'}

\item{biplot}{\code{TRUE} if both latent variables and their coefficients are plotted, \code{FALSE} if only LVs}

\item{ind.spp}{Number of LV coefficients included in the biplot.}

\item{alpha}{A numeric scalar between 0 and 1 that is used to control the relative scaling of the latent variables and their coefficients, when constructing a biplot.}

\item{which.lvs}{Indices of two latent variables to be plotted if number of the latent variables is more than 2. A vector with length of two. Defaults to \code{c(1,2)}.}

\item{jitter}{if \code{TRUE}, jittering is applied on points}

\item{...}{Additional graphical arguments.}
}
\description{
Plot latent variables and their corresponding coefficients (biplot).
}
\details{
Construct a scatter plot of two latent variables, i.e. an ordination plot. If only one latent
variable is in the fitted model, latent variables are plotted against their corresponding row indices.
The latent variables are labeled using the row index of the response matrix y.

Coefficients related to latent variables are plotted in the same figure with the latent
variables if 'biplot=TRUE'. They are labeled using the column names of y. The number
of latent variable coefficients to be plotted can be controlled by ind.spp. An argument alpha
is used to control the relative scaling of the latent variables and their coefficients.
If alpha=0.5, the latent variables and their coefficients are on the same scale.
}
\examples{
\dontrun{
library(mvabund) ## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit GLLVM model
fit <- gllvm(y = y, family = "negative.binomial")
# Ordination plot:
ordplot.gllvm(fit)
# Biplot with 10 species
ordplot.gllvm(fit, biplot = TRUE, ind.spp = 10)
}
}
\references{

}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>, Francis K.C. Hui
}
