% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gllvm.R
\name{plot.gllvm}
\alias{plot.gllvm}
\title{Plot Diagnostics for an gllvm Object}
\usage{
\method{plot}{gllvm}(
  x,
  which = 1:5,
  caption = c("Residuals vs linear predictors", "Normal Q-Q", "Residuals vs row",
    "Residuals vs column", "Scale-Location"),
  var.colors = NULL,
  add.smooth = TRUE,
  envelopes = TRUE,
  reps = 150,
  envelope.col = c("blue", "lightblue"),
  n.plot = NULL,
  spp = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class 'gllvm'.}

\item{which}{if a subset of the plots is required, specify a subset of the numbers 1:5, see caption below. For objects of class "glmmVA" it should be a subset of 1:3.}

\item{caption}{captions to appear above the plots.}

\item{var.colors}{colors for responses, vector with length of number of response variables or 1. Defaults to NULL, when different responses have different colors.}

\item{add.smooth}{logical with default \code{TRUE}. Indicates if a smoother should be added.}

\item{envelopes}{logical, indicating if simulated point-wise confidence interval envelope will be added to Q-Q plot, defaults to \code{TRUE}}

\item{reps}{number of replications when simulating confidence envelopes for normal Q-Q plot}

\item{envelope.col}{colors for envelopes, vector with length of two}

\item{n.plot}{number of species (response variables) to be plotted. Defaults to \code{NULL} when all response variables are plotted. Might be useful when data is very high dimensional.}

\item{spp}{integer vector or \code{NULL}, defaults to \code{NULL}. If provided, residuals are computed only for the selected species (column indices in response matrix).}

\item{...}{additional graphical arguments.}
}
\description{
Five plots (selectable by which) are currently available: a plot of residuals against
linear predictors of fitted values, a Normal Q-Q plot of residuals with a simulated point-wise 95\% confidence interval envelope, residuals against row index and column index and scale location plot.
}
\details{
plot.gllvm is used for model diagnostics. Dunn-Smyth residuals (randomized quantile residuals) (Dunn and Smyth, 1996) are used in plots. Colors indicate different species.
}
\examples{
\dontrun{
# Fit gllvm model with Poisson family
data(microbialdata)
X <- microbialdata$Xenv
y <- microbialdata$Y[, order(colMeans(microbialdata$Y > 0), 
                     decreasing = TRUE)[21:40]]
fit <- gllvm(y, X, formula = ~ pH + Phosp, family = poisson())
# Plot residuals
plot(fit, mfrow = c(3,2))

}
}
\references{
Dunn, P. K., and Smyth, G. K. (1996). Randomized quantile residuals. Journal of Computational and Graphical Statistics, 5, 236-244.

Hui, F. K. C., Taskinen, S., Pledger, S., Foster, S. D., and Warton, D. I. (2015).  Model-based approaches to unconstrained ordination. Methods in Ecology and Evolution, 6:399-411.
}
\seealso{
\code{\link{gllvm}}, \code{\link{residuals.gllvm}}
}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>
}
