% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSR.gllvm.R
\name{plot.predictSR.gllvm}
\alias{plot.predictSR.gllvm}
\title{Plot diagnostics for species richness predicted from a gllvm object}
\usage{
\method{plot}{predictSR.gllvm}(
  x,
  object = NULL,
  which = 1:3,
  caption = c("Residuals vs fitted", "Normal Q-Q", "Observed vs fitted"),
  ...
)
}
\arguments{
\item{x}{an object of class 'predictSR.gllvm'.}

\item{object}{an object of class 'gllvm', defaults to \code{NULL}. Only used if \code{x} is \code{NULL}, in which case \code{"\link{predictSR.gllvm}"} is called on \code{object} first.}

\item{which}{if a subset of the plots is required, specify a subset of the numbers 1:3, see the description above.}

\item{caption}{captions to appear above the plots.}

\item{...}{additional graphical arguments.}
}
\description{
Three plots are available: 1) Dunn-Smyth residuals for species richness against the expected species richness, 2) a Normal Q-Q plot of those residuals, and 3) observed against predicted species richness, with a one-to-one line and, if the \code{predictSR} object includes one, the confidence interval of the prediction.
}
\details{
plot.predictSR.gllvm is used for diagnostics of predicted species richness. Dunn-Smyth residuals (randomized quantile residuals) (Dunn and Smyth, 1996) for species richness are used in the first two plots, see \code{"\link{residuals.predictSR.gllvm}"}.
}
\references{
Dunn, P. K., and Smyth, G. K. (1996). Randomized quantile residuals. Journal of Computational and Graphical Statistics, 5, 236-244.
}
\seealso{
\code{\link{predictSR.gllvm}}, \code{\link{residuals.predictSR.gllvm}}, \code{\link{plot.gllvm}}
}
\author{
Bert van der Veen
}
