% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVP.R
\name{plotVarPartitioning}
\alias{plotVarPartitioning}
\alias{plotVP}
\title{Plot variance partitioning}
\usage{
plotVarPartitioning(
  VP,
  main = "Variance Partitioning",
  xlab = "Response",
  ylab = "Variance proportion",
  legend.text = NULL,
  ...
)
}
\arguments{
\item{VP}{a variance partitioning object for a gllvm produced by function varPartitioning.}

\item{main}{main title}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{legend.text}{a vector of names for the groups, as a default 'groupnames' from varPartitioning. If FALSE, legend not printed.}

\item{...}{additional graphical arguments passed to the barplot function}
}
\description{
Plots the results of variance partitioning of a fitted gllvm.
}
\examples{
\dontrun{
# Calculate variance partitioning for a fitted model 'fit'
VP <- varPartitioning(fit)
# Plot the result of  variance partitioning
plot(VP, col = palette(hcl.colors(5, "viridis")))
}
}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>
}
