% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gllvm.R
\name{print.gllvm}
\alias{print.gllvm}
\title{Print gllvm}
\usage{
\method{print}{gllvm}(x, ...)
}
\arguments{
\item{x}{an object of class 'gllvm'.}

\item{...}{not used.}
}
\description{
Print an object of class 'gllvm'.
}
\examples{
\dontrun{
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y = y, family = "negative.binomial")
# Print:
print(fit)
}
}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>
}
