% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmVA.R
\name{ranef.glmmVA}
\alias{ranef.glmmVA}
\title{Extract random effects from a glmmVA object}
\usage{
\method{ranef}{glmmVA}(object, condVar = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{"glmmVA"}.}

\item{condVar}{logical; if \code{TRUE}, attach conditional variances as an
attribute \code{"condVar"} on the returned object. Defaults to \code{FALSE}.}

\item{...}{not used.}
}
\value{
A vector of random-effect estimates. When \code{condVar = TRUE}, the attribute \code{"condVar"}
  is a vector of conditional variances.
  returned by \code{\link{getPredictErr}}.
}
\description{
Returns the random effects estimate for a model fitted
  with \code{\link{glmmVA}}, optionally with their conditional variances.
}
\seealso{
\code{\link{glmmVA}}, \code{\link{getPredictErr}}
}
\author{
Bert van der Veen
}
