% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmVA.R
\name{ranef}
\alias{ranef}
\title{Generic function for extracting random effects}
\usage{
ranef(object, ...)
}
\arguments{
\item{object}{a fitted model object of class glmmVA.}

\item{...}{additional arguments passed to the method.}
}
\description{
S3 generic dispatched on the model class.
}
