% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gllvm.R
\name{summary.gllvm}
\alias{summary.gllvm}
\title{Summary of fitted gllvm object}
\usage{
\method{summary}{gllvm}(object, ...)
}
\arguments{
\item{object}{An object of class 'gllvm'}

\item{...}{Not used.}
}
\description{
A summary of the fitted gllvm object, including function call, distribution family and model parameters.
}
\examples{
\dontrun{
library(mvabund) ## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit GLLVM model
fit <- gllvm(y = y, family = "negative.binomial")
summary(fit)
}
}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>
}
