% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.gllvm.R
\name{update.gllvm}
\alias{update.gllvm}
\alias{update}
\title{Update and Re-fit a gllvm Model Call}
\usage{
\method{update}{gllvm}(
  object,
  formula = NULL,
  lv.formula = NULL,
  row.eff = NULL,
  eval = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{formula}{interface for column-specific effects.}

\item{lv.formula}{interface for latent variable (reduced-rank) effects.}

\item{row.eff}{interface for row-specific (i.e., same for all columns) effects.}

\item{eval}{default to TRUE, in which case the reformulated model is fitted directly. if FALSE returns the new model call.}

\item{...}{not used.}
}
\description{
Models of class gllvm have three formula interfaces: formula, lv.formula, and row.eff. Each can include fixed and random effects.
}
\details{
Updates the formula of a gllvm object.
}
\author{
Bert van der Veen
}
