% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VP.gllvm.R
\name{varPartitioning.gllvm}
\alias{varPartitioning.gllvm}
\alias{varPartitioning}
\alias{VP}
\title{Calculate variance partitioning}
\usage{
\method{varPartitioning}{gllvm}(
  object,
  group = NULL,
  groupnames = NULL,
  adj.cov = TRUE,
  grouplvs = FALSE
)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{group}{a vector of integers identifying grouping of X covariates, the default is to use model terms formula and lv.formula.}

\item{groupnames}{a vector of strings given as names for the groups defined in group}

\item{adj.cov}{logical, whether or not to adjust co-variation within the group}

\item{grouplvs}{logical, whether or not to group latent variables to one group}
}
\description{
Calculates variance partitioning for gllvm object.
}
\details{
Variance for the linear predictor for response j can be calculated as 

 \deqn{Var(\eta_j) = \sum_k \beta_{jk}^2*var(z_{.k}) + 2 \sum_{(k1=1,...,K-1)} \sum_{(k2=k1+1,...,K)}  \beta_{j(k1)}\beta_{j(k2)} Cov(Z_{.k1},Z_{.k2}) , }
 
where \eqn{z_{.k}} is a vector consisting of predictor/latent variable/row effect etc values for all sampling units i.
If \eqn{z_{.k}}s are not correlated, covariance term is 0 and thus the variance explained of a response j for predictor \eqn{z_{.k}} is given as \eqn{\beta_{jk}^2*var(z_{.k})/Var(\eta_j)}.

In case of correlated predictors, it is advised to group them into a same group. The variance explained is calculated for the correlated group of predictors together and adjusted with the covariance term.
}
\examples{
# Extract subset of the microbial data to be used as an example
data(microbialdata)
X <- microbialdata$Xenv
y <- microbialdata$Y[, order(colMeans(microbialdata$Y > 0), 
                     decreasing = TRUE)[21:40]]
fit <- gllvm(y, X[,1:3], formula = ~ pH + Phosp, family = poisson(), 
             studyDesign = X[,4:5], row.eff = ~(1|Site))
VP <- varPartitioning(fit)
plotVarPartitioning(VP)

}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>
}
