% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFit.gllvm.R
\name{allFit.gllvm}
\alias{allFit.gllvm}
\title{Re-fit a GLLVM with different starting values or optimizers}
\usage{
\method{allFit}{gllvm}(
  object,
  starting.vals = TRUE,
  optimizers = TRUE,
  return.best = TRUE,
  return.table = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class 'gllvm'}

\item{starting.vals}{should the model be re-fitted with different starting values? Defaults to \code{TRUE}.}

\item{optimizers}{should the model be re-fitted with different optimizers? Defaults to \code{TRUE}.}

\item{return.best}{return only the best fitting object (\code{TRUE})? Or all fitted objects \code{FALSE}?}

\item{return.table}{return a table with information of the fitted model objects? Defaults to \code{FALSE}.}

\item{...}{other arguments for a \code{\link{gllvm}} function call.}
}
\description{
The algorithm used is sensitive to the initial values, so that a single model fit might not always provide the best parameter estimates. Frequent re-fitting might be required to ensure the best solution has been found. This function re-fits GLLVM objects with different starting values or optimizers, for convenience, to ensure an optimal solution.
}
\examples{
\dontrun{
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y = y, family = poisson())
bestFit <- allFit(fit)
}
}
\seealso{
\code{\link{gllvm}}.
}
\author{
Bert van der Veen
}
