% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.gllvm.R
\name{coef.gllvm}
\alias{coef.gllvm}
\title{Extract gllvm coefficients}
\usage{
\method{coef}{gllvm}(object, ...)
}
\arguments{
\item{object}{An object of class 'gllvm'}

\item{...}{Not used.}
}
\description{
Extracts model coefficients from 'gllvm' objects.
}
\examples{
\dontrun{
library(mvabund) ## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit GLLVM model
fit <- gllvm(y = y, family = "negative.binomial")
# Coefficients
coef(fit)
}
}
\author{
David Warton, Jenni Niku <jenni.m.e.niku@jyu.fi>
}
