% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmVA.R
\name{glmmVA}
\alias{glmmVA}
\title{Fitting Generalized Linear Mixed-Effects Models with Variational Approximation}
\usage{
glmmVA(formula, data, family, ...)
}
\arguments{
\item{formula}{a formula object describing both the fixed- and random-effects part of the model. A response should be present on the left-hand side of the operator, and otherwise passed as a named 'y' argument to the function. Random effects are written in \code{lme4}-style. Some structured random effects are supported, see \code{"\link{gllvm}"} for more information.}

\item{data}{an optional data frame containing the variables named in formula.}

\item{family}{a family as supported by the \code{"\link{gllvm}"} function.}

\item{...}{other arguments passed onto the \code{"\link{gllvm}"} function.}
}
\value{
An object of class "glmmVA" that inherits from the "gllvm" class.
}
\description{
Wraps the \code{"\link{gllvm}"} function to fit a (univaroate) generalized linear mixed-effects model (GLMM) in a more familiar syntax. 
Both fixed and random effects are specified in \code{lme4}-style via the formula argument.
}
\examples{
data(eSpider)
data <- cbind(data.frame(y = c(eSpider$abund[eSpider$nonNA,]),
                         species = factor(rep(1:ncol(eSpider$abund), each = length(eSpider$nonNA))),
                         site = factor(rep(1:length(eSpider$nonNA), ncol(eSpider$abund)))),
                         do.call(rbind, replicate(ncol(eSpider$abund), scale(eSpider$X[eSpider$nonNA,]), simplify = FALSE)))
# Example 1: crossed random slope effects                          
model <- glmmVA(y~species + ConWate + ConHumu + (0+ConHumu|species) + (0+ConWate|species), family = "poisson", data = data)

# Example 2: correlated random slopes
model <- glmmVA(y~species + ConWate + ConHumu + (0+ConWate+ConHumu|species), family = "poisson", data = data)

# Example 3: 
}
\seealso{
\code{\link{gllvm}}
}
\author{
Bert van der Veen
}
