% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmVA.R
\name{glmmVA}
\alias{glmmVA}
\title{Fitting Generalized Linear Mixed-Effects Models with Variational Approximation}
\usage{
glmmVA(
  formula,
  data,
  family,
  control = list(reltol = 1e-10, optimizer = "optim", max.iter = 6000, maxit = 6000,
    optim.method = "BFGS"),
  control.va = list(Ar.struc = "unstructured", diag.iter = 0, Lambda.start = 0.3),
  ...
)
}
\arguments{
\item{formula}{a formula object describing both the fixed- and random-effects part of the model. A response should be present on the left-hand side of the operator, and otherwise passed as a named 'y' argument to the function. Random effects are written in \code{lme4}-style. Some structured random effects are supported, see \code{"\link{gllvm}"} for more information.}

\item{data}{an optional data frame containing the variables named in formula.}

\item{family}{a family as supported by the \code{"\link{gllvm}"} function.}

\item{control}{A list with the following arguments controlling the optimization:
\describe{
 \item{\emph{reltol}: }{ convergence criteria for log-likelihood, defaults to 1e-10.}
 \item{\emph{optimizer}: }{the log-likelihood can be optimized using \code{"\link{optim}"} (default) or \code{"\link{nlminb}"}.}
 \item{\emph{max.iter}: }{ maximum number of iterations for \code{optimizer = "nlminb"}, defaults to 6000.}
 \item{\emph{maxit}: }{ maximum number of iterations for optimizer, defaults to 6000.}
 \item{\emph{optim.method}: }{ optimization method to be used if optimizer is \code{"\link{optim}"}. Defaults to \code{"BFGS"}, but to \code{"L-BFGS-B"} for Tweedie family due the limited-memory use.}
}}

\item{control.va}{A list with the following arguments controlling the variational approximation method:
\describe{
 \item{\emph{Ar.struc}: }{ covariance structure of VA distributions, "unstructured" or "diagonal". Defaults to "unstructured". "Unstructured" meaning block diagonal for ordinary random effects, and fully unstructured for structured random effects (such as "corExp").}
 \item{\emph{diag.iter}: }{ non-negative integer which can sometimes be used to speed up the updating of variational (covariance) parameters in VA method. Can sometimes improve the accuracy. Either 0 or 1. Defaults to 0.}
 \item{\emph{Lambda.start}: }{ starting value for variances in VA distributions. Defaults to 0.3.}
}}

\item{...}{other arguments passed onto the \code{"\link{gllvm}"} function.}

\item{control.start}{A list of arguments controlling the starting values. See \code{"\link{gllvm}"} for details.}
}
\value{
An object of class "glmmVA" that inherits from the "gllvm" class.
}
\description{
Wraps the \code{"\link{gllvm}"} function to fit a (univaroate) generalized linear mixed-effects model (GLMM) in a more familiar syntax. 
Both fixed and random effects are specified in \code{lme4}-style via the formula argument.
}
\examples{
data(eSpider)
data <- cbind(data.frame(y = c(eSpider$abund[eSpider$nonNA,]),
                         species = factor(rep(1:ncol(eSpider$abund), each = length(eSpider$nonNA))),
                         site = factor(rep(1:length(eSpider$nonNA), ncol(eSpider$abund)))),
                         do.call(rbind, replicate(ncol(eSpider$abund), scale(eSpider$X[eSpider$nonNA,]), simplify = FALSE)))
# Example 1: crossed random slope effects                          
model <- glmmVA(y~species + ConWate + ConHumu + (0+ConHumu|species) + (0+ConWate|species), family = "poisson", data = data)

# Example 2: correlated random slopes
model1 <- glmmVA(y~species + ConWate + ConHumu + (0+ConWate+ConHumu|species), family = "poisson", data = data)

}
\seealso{
\code{\link{gllvm}}
}
\author{
Bert van der Veen
}
