% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodnessOfFit.gllvm.R
\name{goodnessOfFit}
\alias{goodnessOfFit}
\title{Goodness of fit measures for a gllvm}
\usage{
goodnessOfFit(
  object = NULL,
  y = NULL,
  pred = NULL,
  measure = c("cor", "RMSE", "MAE", "MARNE"),
  species = FALSE
)
}
\arguments{
\item{object}{an object of class 'gllvm', to calculate goodness of a model fit.}

\item{y}{a response matrix of new observations}

\item{pred}{predicted values for response matrix y if you want to calculate prediction accuracy for new values. Note that for ordinal model, you need to give the predicted classes.}

\item{measure}{a goodness-of-fit measure to be calculated. Options are \code{"cor"} (correlation between observed and predicted values), \code{"scor"} (Spearman correlation between observed and predicted values), \code{"RMSE"} (root mean squared error of prediction), \code{"MAE"} (Mean Absolute Error), \code{"MARNE"} (Mean Absolute Range Normalized Error), \code{"TjurR2"} (Tjur's R2 measure, only for binary data), \code{"R2"} (R-squared as the square of the correlation), "AUC", \code{"sR2"} (R-squared as the square of the spearman correlation). Likelihood based pseudo R2 meaures \code{"NagelkerkeR2"}, \code{"McFaddenR2"}, \code{"CoxSnellR2"} can be calculated currently only for training data to measure the model's (in sample) goodness of fit for full data, not response specific.}

\item{species}{logical, if \code{TRUE}, goodness-of-fit measures are calculated for each species separately. If FALSE,  goodness-of-fit measures are calculated for all species together.}
}
\description{
Several goodness-of-fit measure are currently available and can be calculated for a gllvm model fit and predicted values.
}
\details{
goodnessOfFit is used for evaluating the goodness-of-fit of a model or predictions. Available goodness-of-fit measures are correlation, RMSE, MARNE, and R2 measures. Definitions are below.
Denote an observed response j (species) at sample i, \eqn{i=1,...,n}, as \eqn{y_{ij}}, and predicted value as \eqn{\hat y_{ij}}.

\deqn{RMSE(\boldsymbol{y_{j}}, \boldsymbol{\hat y_{j}}) =  \sqrt{\frac{1}{n}\Sigma_{i=1}^{n} {(y_{ij} - \hat y_{ij})^2}} }

\deqn{MAE(\boldsymbol{y_{j}}, \boldsymbol{\hat y_{j}}) =  \frac{1}{n}\Sigma_{i=1}^{n} |y_{ij} - \hat y_{ij}| }

\deqn{MARNE(\boldsymbol{y_{j}}, \boldsymbol{\hat y_{j}}) =  \frac{1}{n}\Sigma_{i=1}^{n} \frac{|y_{ij} - \hat y_{ij}|}{max(\boldsymbol{y_{j}}) - min(\boldsymbol{y_{j}})} }

\deqn{\text{Tjur's } R2(\boldsymbol{y_{j}}, \boldsymbol{\hat y_{j}}) =  \frac{1}{n_1}\Sigma \hat y_{ij}\boldsymbol{1}_{y=1}(y_{ij}) - \frac{1}{n_0}\Sigma \hat y_{ij}\boldsymbol{1}_{y=0}(y_{ij}) }

Pseudo R-squared measures based on likelihood of the evaluated model \eqn{L_M} and null model \eqn{L_M} with intercept only and total number of observations N:

 \deqn{\text{McFadden } R2 = 1- (\frac{log L_M}{log L_0}) }

 \deqn{\text{Cox and Snell } R2 = 1- (\frac{L_0}{L_M})^{\frac{2}{N}} }

 \deqn{\text{Nagelkerke } R2 =  \frac{1- (\frac{L_0}{L_M})^{\frac{2}{N}} }{1- (L_0)^{\frac{2}{N}}} }
}
\examples{
\dontrun{
# Fit gllvm model with Poisson family
data(microbialdata)
X <- microbialdata$Xenv
y <- microbialdata$Y[, order(colMeans(microbialdata$Y > 0), 
                     decreasing = TRUE)[21:40]]
fit <- gllvm(y, X, formula = ~ pH + Phosp, family = poisson())
# Calculate metrics
goodnessOfFit(object = fit, measure = c("cor", "RMSE"))

}
}
\seealso{
\code{\link{gllvm}}, \code{\link{predict.gllvm}}
}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>
}
