% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optiplot.gllvm.quadratic.R
\name{optiplot.gllvm.quadratic}
\alias{optiplot.gllvm.quadratic}
\alias{optiplot}
\title{Plot species optima for latent variables from a gllvm}
\usage{
\method{optiplot}{gllvm.quadratic}(
  object,
  ind.spp = NULL,
  alpha = 0.5,
  main = NULL,
  which.lvs = NULL,
  s.colors = 1,
  cex.spp = 0.7,
  opt.ranges = FALSE,
  type = "response",
  intercept = TRUE,
  legend = FALSE,
  scale = FALSE,
  predict.region = FALSE,
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{ind.spp}{the number of response variables (usually, species) to include on the biplot (sorted by varianve explained). The default is all.}

\item{alpha}{a numeric scalar between 0 and 1 that is used to control the relative scaling of the latent variables and their coefficients, when constructing a biplot.}

\item{main}{main title.}

\item{which.lvs}{indices of two latent variables to be plotted if number of the latent variables is more than 2. A vector with length of two. Defaults to \code{c(1,2)}.}

\item{s.colors}{colors for sites}

\item{cex.spp}{size of species labels in biplot}

\item{opt.ranges}{Only for 2D plots, efaults to FALSE. If "statistical", plots statistical uncertainties for species optima. If "environmental" plots preicted environmental ranges}

\item{type}{Can be used to predict on the response or link scale. Default is response (except for ordinal, for which the only option is "link").}

\item{intercept}{Can be used to include species-intercepts in the plot. Default is TRUE}

\item{legend}{when \code{TRUE} adds legend in the topleft corner of the plot, instead of species names in the plot}

\item{scale}{For 2D plots, either "FALSE",species" or "sites" to scale optima or site scores by the ratio variance explained. Alternatively can be "tolerances" to scale optima by tolerances and site scores by average tolerances per latent variable.}

\item{predict.region}{logical, if \code{TRUE} prediction regions for the predicted latent variables are plotted, defaults to \code{FALSE}.}

\item{level}{level for prediction regions of sites and optima.}

\item{...\tadditional}{graphical arguments.}
}
\description{
Plots latent variables and their corresponding species optima.
}
\details{
Function constructs a plot of species optima.

Plots species optima in one or two dimensions, potentially with predicted environmental ranges.
}
\examples{
#'## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
fit <- gllvm.quadratic(y, family = poisson())
# Ordination plot:
optiplot(fit)
# Optiplot with 10 species
optiplot(fit, ind.spp = 10)

}
\author{
Bert van der Veen
}
