% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSR.gllvm.R
\name{predictPairwise.gllvm}
\alias{predictPairwise.gllvm}
\alias{predictPairwise}
\title{Predict co-occurrence from a gllvm object}
\usage{
\method{predictPairwise}{gllvm}(
  object,
  spp = NULL,
  se.fit = FALSE,
  alpha = 0.95,
  seed = 42,
  ...
)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{spp}{a vector of length 2, defaults to NULL. If NULL, returns the probability of co-occurrence for all species in the data.}

\item{se.fit}{logical, defaults to \code{FALSE}. If set to \code{TRUE}, returns a simulated confidence interval of the prediction.}

\item{alpha}{numeric between 0 and 1, defaults to 0.95. Confidence level of the prediction.}

\item{seed}{numeric, defaults to 42. Seed for the simulation of the confidence interval.}

\item{...}{arguments passed to \code{"\link{predict.gllvm}"}.}
}
\value{
A matrix of size np(p-1)/2 by 2.
}
\description{
Obtains the probability of co-occurrnece for species from a fitted generalized linear latent variable model object.
}
\details{
The probability of co-occurrence is simply given by the joint probability of getting two presences for any pair of species. As such,
it is simply calculated as the product of the probability of occurrence for any pair of species, so that this function mostly serves
as convenience wrapper around \code{"\link{predict.gllvm}"}.
}
\examples{

\donttest{
# Load a dataset from the mvabund package
data(spider, package = "mvabund")
y <- as.matrix(spider$abund)
X <- scale(spider$x)
# Fit gllvm model
fit <- gllvm(y = y, X, formula = ~soil.dry, family = poisson())
# fitted values
newX = data.frame(soil.dry = seq(min(X[,"soil.dry"]),
                 max(X[,"soil.dry"]), length.out=100))
predPair <- predictPairwise(fit, spp = c(1,2), newX = newX, level = 0)
plot(x = newX$soil.dry, y = predPair$prob, type = "l", 
    xlab = "Soil dry moisture content", 
    ylab = "Probability of co-occurrence", ylim = c(0,1))

}

}
\seealso{
\code{\link{predict.gllvm}}, \code{\link{predictSR.gllvm}}
}
\author{
Bert van der Veen
}
