% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSR.gllvm.R
\name{predictSR.gllvm}
\alias{predictSR.gllvm}
\alias{predictSR}
\title{Predict species richness from a gllvm object}
\usage{
\method{predictSR}{gllvm}(object, SR = NULL, se.fit = FALSE, alpha = 0.95, seed = 42, ...)
}
\arguments{
\item{object}{an object of class 'gllvm'.}

\item{SR}{integer, defaults to NULL. If omitted, returns a prediction for all integers up to the number of observed species.}

\item{se.fit}{logical, defaults to \code{FALSE}. If set to \code{TRUE}, returns a simulated confidence interval of the prediction.}

\item{alpha}{numeric between 0 and 1, defaults to 0.95. Confidence level of the prediction.}

\item{seed}{numeric, defaults to 42. Seed for the simulation of the confidence interval.}

\item{...}{arguments passed to \code{"\link{predict.gllvm}"}.}
}
\value{
A matrix of size sites by length(SR).
}
\description{
Obtains predictions from a fitted generalized linear latent variable model object.
}
\examples{
\donttest{
# Load a dataset from the mvabund package
data(spider, package = "mvabund")
y <- as.matrix(spider$abund)
X <- scale(spider$x)
# Fit gllvm model
fit <- gllvm(y = y, X, formula = ~soil.dry, family = poisson())
# Fitted values
newX = data.frame(soil.dry = seq(min(X[,"soil.dry"]),max(X[,"soil.dry"]),length.out=100))
predSR <- predictSR(fit, newX = newX, level = 0)

# Visualize the results
par(mfrow = c(4,4))
for(i in 0:ncol(fit$y)){
plot(x = newX$soil.dry, y = predSR[,i+1], xlab = "Soil dry matter content", ylab = bquote(p(SR == .(i))), type = "l", ylim = c(0,1))
}

}

}
\seealso{
\code{\link{predict.gllvm}}
}
\author{
Bert van der Veen
}
