% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.gllvm.R
\name{residuals.gllvm}
\alias{residuals.gllvm}
\title{Dunn-Smyth -residuals for GLLVM model}
\usage{
\method{residuals}{gllvm}(object, plot = TRUE, by.index = NULL,
  xlim = NULL, ...)
}
\arguments{
\item{object}{An object of class 'gllvm'}

\item{plot}{Logical, construct a plot of residuals. Defaults to \code{FALSE}}

\item{by.index}{If \code{NULL} (default), residuals are plotted by linear predictors. If \code{by.index="site"}, residuals are plotted by row index. If \code{by.index="spp"} residuals are plotted by column index.}

\item{...}{Additional graphical arguments.}
}
\value{
A list containing \code{residuals} which is a matrix of residuals and \code{linpred} which is a matrix of linear predictors.
}
\description{
Calculates Dunn-Smyth -residuals for GLLVM model and plots them.
}
\details{
Computes Dunn-Smyth residuals or randomized quantile residuals (Dunn and Smyth, 1996) for GLLVM model.
}
\examples{
\dontrun{
library(mvabund) ## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit GLLVM model
fit <- gllvm(y = y, family = "negative.binomial")
# residuals saved and plotted by linear predictors
res <- residuals(fit, plot = TRUE, by.index = NULL)
# residuals plotted by site index
res <- residuals(fit, plot = TRUE, by.index = "site")
# residuals plotted by column index
res <- residuals(fit, plot = TRUE, by.index = "spp")
}
}
\references{
Dunn, P. K., and Smyth, G. K. (1996). Randomized quantile residuals. Journal of Computational and Graphical Statistics, 5, 236-244.
}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>
}
