% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSR.gllvm.R
\name{residuals.predictSR.gllvm}
\alias{residuals.predictSR.gllvm}
\title{Dunn-Smyth residuals for species richness from a gllvm object}
\usage{
\method{residuals}{predictSR.gllvm}(object, model, ...)
}
\arguments{
\item{object}{object returned by \code{"\link{predictSR.gllvm}"}}

\item{model}{an object of class 'gllvm'.}

\item{...}{not used.}
}
\value{
A list of length 2 with 1) the fitted and 2) the dunn-smyth residuals.
}
\description{
Calculates Dunn-Smyth residuals for species richness.
}
\details{
See \code{"\link{residuals.gllvm}"} for details.
}
\seealso{
\code{\link{predict.gllvm}}, \code{\link{predictSR.gllvm}}, \code{\link{residuals.gllvm}}
}
\author{
Bert van der Veen
}
