% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSR.gllvm.R
\name{residuals.predictSR.gllvm}
\alias{residuals.predictSR.gllvm}
\title{Dunn-Smyth residuals for species richness from a gllvm object}
\usage{
\method{residuals}{predictSR.gllvm}(object, ...)
}
\arguments{
\item{object}{object returned by \code{"\link{predictSR.gllvm}"}}

\item{...}{not used.}
}
\value{
A list of length 3 with the fitted, the dunn-smyth residuals, and the observed species richness.
}
\description{
Calculates Dunn-Smyth residuals for species richness.
}
\details{
See \code{"\link{residuals.gllvm}"} for details.
}
\seealso{
\code{\link{predict.gllvm}}, \code{\link{predictSR.gllvm}}, \code{\link{plot.predictSR.gllvm}}, \code{\link{residuals.gllvm}}
}
\author{
Bert van der Veen
}
