\name{sim_bacteria}
\docType{data}
\alias{sim_bacteria}
\title{Simulated bacteria data}
\usage{
data(sim_bacteria)
}
\format{
\describe{
\item{y}{A data frame with abundances of 985 species measured at 56 sites}
\item{X}{Environmental variables Temperature and pH}
}}
\description{
Simulated dataset of size 56 x 985 that can be used to illustrate the methods in Section 5.1 of Niku et al., (2017). 
}
\references{
Niku, J., Warton,  D. I., Hui, F. K. C., and Taskinen, S. (2017). Generalized Linear Latent Variable Models for Multivariate Count and Biomass Data in Ecology. Journal of Agricultural, Biological, and Environmental Statistics, 22, 498-522.
}