% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gllvm.R
\name{summary.gllvm}
\alias{summary.gllvm}
\title{Summarizing gllvm model fits}
\usage{
\method{summary}{gllvm}(object, ...)
}
\arguments{
\item{object}{an object of class 'gllvm'}

\item{...}{not used.}
}
\description{
A summary of the fitted 'gllvm' object, including function call, distribution family and model parameters.
}
\examples{
\dontrun{
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y = y, family = "poisson")
summary(fit)
}
}
\author{
Jenni Niku <jenni.m.e.niku@jyu.fi>
}
